Python_software_engineering/Raspberry Pi/requirements.txt
requests
pygame
keyboard
numpy
pyzbar
matplotlib
wheel
PILLOW